+++ beaker/client.py
@@ -4,7 +4,7 @@ import time
 import urllib.parse
 from contextlib import contextmanager
 from pathlib import Path
-from typing import Any, Dict, Generator, List, Optional, Union
+from typing import Any, Deque, Dict, Generator, List, Optional, Union
 
 import docker
 import requests
@@ -15,6 +15,7 @@ from rich.progress import (
     BarColumn,
     FileSizeColumn,
     Progress,
+    ProgressColumn,
     SpinnerColumn,
     TaskID,
     TimeElapsedColumn,
@@ -171,6 +172,7 @@ class ServiceClient:
             default_headers = {
                 "Authorization": f"Bearer {token or self.config.user_token}",
                 "Content-Type": "application/json",
+                "User-Agent": f"beaker-py v{VERSION}",
             }
             if headers is not None:
                 default_headers.update(headers)
@@ -263,8 +265,6 @@ class DatasetClient(ServiceClient):
     HEADER_UPLOAD_LENGTH = "Upload-Length"
     HEADER_UPLOAD_OFFSET = "Upload-Offset"
     HEADER_DIGEST = "Digest"
-    HEADER_USER_AGENT = "User-Agent"
-    USER_AGENT = f"beaker-py v{VERSION}"
 
     SHA256 = "SHA256"
 
@@ -277,8 +277,8 @@ class DatasetClient(ServiceClient):
         target: Optional[PathOrStr] = None,
         workspace: Optional[str] = None,
         force: bool = False,
-        quiet: bool = False,
         max_workers: int = 8,
+        quiet: bool = False,
     ) -> Dataset:
         """
         Create a dataset with the source file(s).
@@ -291,8 +291,8 @@ class DatasetClient(ServiceClient):
         :param workspace: The workspace to upload the dataset to. If not specified,
             :data:`Beaker.config.default_workspace <beaker.Config.default_workspace>` is used.
         :param force: If ``True`` and a dataset by the given name already exists, it will be overwritten.
-        :param quiet: If ``True``, progress won't be displayed.
         :param max_workers: The maximum number of thread pool workers to use to upload files concurrently.
+        :param quiet: If ``True``, progress won't be displayed.
 
         :raises DatasetConflict: If a dataset by that name already exists and ``force=False``.
         :raises WorkspaceNotFound: If the workspace doesn't exist.
@@ -333,7 +333,7 @@ class DatasetClient(ServiceClient):
         assert dataset_info.storage is not None
 
         # Upload the file(s).
-        self.sync_source(dataset_info, source, target, quiet=quiet, max_workers=max_workers)
+        self._sync_source(dataset_info, source, target, quiet=quiet, max_workers=max_workers)
 
         # Commit the dataset.
         self.request(
@@ -344,6 +344,90 @@ class DatasetClient(ServiceClient):
 
         # Return info about the dataset.
         return self.get(dataset_info.id)
+
+    def fetch(
+        self,
+        dataset: str,
+        target: Optional[PathOrStr] = None,
+        force: bool = False,
+        max_workers: int = 8,
+        quiet: bool = False,
+    ):
+        """
+        Download a dataset.
+
+        :param dataset: The dataset ID, full name, or object.
+        :param target: The target path to fetched data. Defaults to ``Path(.)``.
+        :param max_workers: The maximum number of thread pool workers to use to download files concurrently.
+        :param force: If ``True``, existing local files will be overwritten.
+        :param quiet: If ``True``, progress won't be displayed.
+
+        :raises DatasetNotFound: If the dataset can't be found.
+        :raises FileExistsError: If ``force=False`` and an existing local file clashes with a file
+            in the Beaker dataset.
+        :raises HTTPError: Any other HTTP exception that can occur.
+        """
+        dataset: Dataset = self.get(dataset)
+        assert dataset.storage is not None
+
+        storage_info = DatasetStorageInfo.from_json(
+            self.request(
+                f"datasets/{dataset.storage.id}",
+                method="GET",
+                token=dataset.storage.token,
+                base_url=dataset.storage.address,
+            ).json()
+        )
+
+        target: Path = Path(target or Path("."))
+        target.mkdir(exist_ok=True, parents=True)
+
+        total_bytes_to_download: Optional[int] = None
+        total_downloaded: int = 0
+        progress_columns: List[Union[str, ProgressColumn]]
+        if storage_info.size is not None and storage_info.size.final:
+            total_bytes_to_download = storage_info.size.bytes
+            progress_columns = [
+                "[progress.description]{task.description}",
+                BarColumn(),
+                "[progress.percentage]{task.percentage:>3.0f}%",
+                TimeElapsedColumn(),
+                TimeRemainingColumn(),
+                DownloadUploadColumn(),
+            ]
+        else:
+            progress_columns = [
+                "[progress.description]{task.description}",
+                SpinnerColumn(),
+                TimeElapsedColumn(),
+                DownloadUploadColumn(),
+            ]
+
+        with Progress(*progress_columns, disable=quiet) as progress:
+            bytes_task = progress.add_task("Downloading dataset")
+            if total_bytes_to_download is not None:
+                progress.update(bytes_task, total=total_bytes_to_download)
+
+            import concurrent.futures
+
+            with concurrent.futures.ThreadPoolExecutor(max_workers=max_workers) as executor:
+                download_futures = []
+                for file_info in self._iter_files(dataset.storage):
+                    if total_bytes_to_download is None:
+                        progress.update(bytes_task, total=total_downloaded + file_info.size + 1)
+                    target_path = target / Path(file_info.path)
+                    if not force and target_path.exists():
+                        raise FileExistsError(file_info.path)
+                    future = executor.submit(
+                        self._download_file, file_info, target_path, progress, bytes_task
+                    )
+                    download_futures.append(future)
+
+                for future in concurrent.futures.as_completed(download_futures):
+                    total_downloaded += future.result()
+
+            if total_bytes_to_download is None:
+                progress.update(bytes_task, total=total_downloaded, completed=total_downloaded)
 
     def get(self, dataset: str) -> Dataset:
         """
@@ -384,11 +468,10 @@ class DatasetClient(ServiceClient):
             f"*full* name of the dataset (with the account prefix, e.g. 'username/dataset_name')"
         )
 
-    def sync_source(
+    def _sync_source(
         self,
         dataset: Dataset,
-        source,
-        PathOrStr,
+        source: PathOrStr,
         target: Optional[PathOrStr] = None,
         quiet: bool = False,
         max_workers: int = 8,
@@ -399,6 +482,7 @@ class DatasetClient(ServiceClient):
             "[progress.description]{task.description}",
             BarColumn(),
             "[progress.percentage]{task.percentage:>3.0f}%",
+            TimeElapsedColumn(),
             TimeRemainingColumn(),
             DownloadUploadColumn(),
             disable=quiet,
@@ -430,7 +514,7 @@ class DatasetClient(ServiceClient):
                 progress.update(bytes_task, total=total_bytes)
 
                 # Now upload.
-                with concurrent.futures.ThreadPoolExecutor(max_workers=8) as executor:
+                with concurrent.futures.ThreadPoolExecutor(max_workers=max_workers) as executor:
                     # Dispatch tasks to thread pool executor.
                     future_to_path = {}
                     for path, size in path_to_size.items():
@@ -487,7 +571,6 @@ class DatasetClient(ServiceClient):
                     method="POST",
                     token=dataset.storage.token,
                     base_url=dataset.storage.address,
-                    headers={self.HEADER_USER_AGENT: self.USER_AGENT},
                 )
                 upload_id = response.headers[self.HEADER_UPLOAD_ID]
 
@@ -503,7 +586,6 @@ class DatasetClient(ServiceClient):
                         token=dataset.storage.token,
                         base_url=dataset.storage.address,
                         headers={
-                            self.HEADER_USER_AGENT: self.USER_AGENT,
                             self.HEADER_UPLOAD_LENGTH: str(size),
                             self.HEADER_UPLOAD_OFFSET: str(written),
                         },
@@ -520,18 +602,13 @@ class DatasetClient(ServiceClient):
 
                 body = None
 
-            headers = {
-                self.HEADER_USER_AGENT: self.USER_AGENT,
-            }
-            if digest:
-                headers[self.HEADER_DIGEST] = self._encode_digest(digest)
             self.request(
                 f"datasets/{dataset.storage.id}/files/{str(target)}",
                 method="PUT",
                 data=body,
                 token=dataset.storage.token,
                 base_url=dataset.storage.address,
-                headers=headers,
+                headers=None if not digest else {self.HEADER_DIGEST: self._encode_digest(digest)},
                 stream=body is not None,
             )
 
@@ -547,6 +624,56 @@ class DatasetClient(ServiceClient):
 
         _, encoded = encoded_digest.split(" ", 2)
         return base64.standard_b64decode(encoded)
+
+    def _iter_files(self, storage: DatasetStorage) -> Generator[FileInfo, None, None]:
+        from collections import deque
+
+        files: Deque[FileInfo] = deque([])
+        last_request: bool = False
+        cursor: Optional[str] = ""
+        while files or not last_request:
+            if files:
+                yield files.popleft()
+            else:
+                manifest = DatasetManifest.from_json(
+                    self.request(
+                        f"datasets/{storage.id}/manifest",
+                        method="GET",
+                        token=storage.token,
+                        base_url=storage.address,
+                        query={"cursor": cursor, "path": "", "url": True},
+                    ).json()
+                )
+                files.extend(manifest.files)
+                cursor = manifest.cursor
+                if not cursor:
+                    last_request = True
+
+    def _download_file(
+        self, file_info: FileInfo, target_path: Path, progress: Progress, task_id: TaskID
+    ) -> int:
+        import tempfile
+
+        total_bytes = 0
+        target_dir = target_path.parent
+        target_dir.mkdir(exist_ok=True, parents=True)
+        with self._session_with_backoff() as session:
+            response = session.get(file_info.url, stream=True)
+            response.raise_for_status()
+            tmp_target = tempfile.NamedTemporaryFile(
+                "w+b", dir=target_dir, delete=False, suffix=".tmp"
+            )
+            try:
+                for chunk in response.iter_content(chunk_size=1024):
+                    total_bytes += len(chunk)
+                    tmp_target.write(chunk)
+                    progress.update(task_id, advance=len(chunk))
+                os.replace(tmp_target.name, target_path)
+            finally:
+                tmp_target.close()
+                if os.path.exists(tmp_target.name):
+                    os.remove(tmp_target.name)
+        return total_bytes
 
 
 class ImageClient(ServiceClient):